Support for command-line specified config files.
If there is string `@path.conf' in the command line, read command line
options from that file and append it to options that are read from
configs. If there is string `@@path.conf', read command line options
only from that file, not from common configs. If there is string `@@',
do not read any configs.